Move BaseTimetable to airflow.sdk.bases
