+++ task-sdk/src/airflow/sdk/definitions/timetables/trigger.py
@@ -21,9 +21,9 @@ from typing import TYPE_CHECKING
 
 import attrs
 
+from airflow.sdk.bases.timetable import BaseTimetable
 from airflow.sdk.definitions.timetables._cron import CronMixin
 from airflow.sdk.definitions.timetables._delta import DeltaMixin
-from airflow.sdk.definitions.timetables.base import BaseTimetable
 
 if TYPE_CHECKING:
     from dateutil.relativedelta import relativedelta
